Signin component destructured in 2 smaller components
